reorganize as a package
