+++ featurevis/__init__.py
@@ -1,0 +1,1 @@
+from featurevis.core import gradient_ascent
