+++ featurevis/exceptions.py
@@ -1,0 +1,3 @@
+class FeatureVisException(Exception):
+    """ Catch-all exception for errors generated during feature visualization."""
+    pass
